crypto/ed25519: Make the batch verification benchmark more realistic
The old benchmark measured:
 * `sigsCount` GenPrivKey + Sign
 * `sigsCount` BatchVerifier.Add
 * `b.N/sigsCount` BatchVerifier.Verify (majority of the runtime).

This is all sorts of nonsensical, especially give that there can be a
non-trivial amount of overhead in `BatchVerifier.Add`.

The rewritten benchmark measures:
 * `b.N/sigsCount` NewBatchVerifier
 * `b.N/sigsCount * sigsCount` BatchVerifier.Add
 * `b.N/sigsCount` BatchVerifier.Verify

This is far more sensible as it better reflects the per-signature
combined cost of instantiating the batch verifier, adding a signature
to the batch, and the verify.