crypto: Refactor the batch verification API
Having to redo quite a bit of computation in the event of a batch
failure if the caller is actually interested in which signature failed,
is rather sub-optimal.  Change the batch verification interface to
expose a one-shot batch verify + fallback call, so that sensible
libraries can handle this case faster.

This commit also leverages the failure information so that validation
will only ever call one of `verifyCommitBatch` or `verifyCommitSingle`.