crypto/sr25519: Switch the sr25519 provider to curve25519-voi
Switch the sr25519 implementation to curve25519-voi for better
performance, and code quality.

Performance comparisions should still be taken with a grain of salt
for the following reasons:

 * curve25519-voi's sr25519 support can use more optimization.
 * go-schnorrkel cuts corners in places by:
   * Not doing delinearization at all when verifying batches
   * Not using the secret key nonce at all when signing.
   * Not sampling random scalars at all when verifying batches,
     unless the import is bumped.

WARNING: This is a breaking change as the original tendermint sr25519
support expands the MiniSecretKey twice, while this implementation
only does it once.