Fix and clarify breaks from select cases.
File list:

    % staticcheck -checks SA4011 ./... | cut -d: -f-2

This change is intended to be semantics-preserving, at least for the intended
semantics of the code. It is NOT strictly behaviour-preserving, however.
Specifically: On some inputs, these loops may have run longer than they were
intended to, prior to this change.

There is one case where I was not sure how to resolve the ambiguity.
This is flagged as FIXME in the commit.