Move order-dependent tests into subtests.
This test relies on block indexing to happen before event indexing.  That
worked, because the test harness executes test functions by default in the
order they are written in the file. However, that's not necessarily obvious to
the reader, and hides the intent.

To fix that, move these two tests into subtest functions. This does not change
the behaviour, but makes the ordering constraint more obvious.