fix:state sync locks when trying to retrieve AppHash
Inside state sync AppHash lookup, we try to download
light block for future heights (`height+2`, `height+2`).
In some scenarios this fail (correctly), because this block
does not exist yet.

In case of error, error handling in `lightBlockFromPrimaryAtHeight()`
assumes the primary is "broken", and tries to find new primary
(`findNewPrimary()`) that will work correctly. Unfortunately,
`findNewPrimary()` was trying to retrieve most recent block,
effectively missing the fact that error happened due to height that
does not exist yet (eg.  `height+2`).

In effect, `findNewPrimary()` returned most recent block, which
in turn was sent to AppHash() and was NOT what was expected.