build: fix the image to work with golangci-lint (#201)
In #195 I updated the Go version, and also swapped from Debian to Alpine for
the base image. This was OK for the build, but it turns out we also use the
same image to run golangci-lint. That action uses a binary for a different
architecture, which fails on this image.

So:
- Keep the Go version, but revert the base image.
- Split image layers to make rebuilds faster.
- Add version labels as arguments.
- Add some documentation to the workflow configs.