Prepare a new compatibility release v0.6.6. (#203)
Revert the breaking API and package structure changes on master, and
prepare to tag a new release compatible with the previous v0.6.4 so that
normal upgrades with go get will function correctly.

The API changes in tag v0.6.5 have been preserved in the new package-reorg
branch, so that any existing dependents will continue  to work. 

Other changes since v0.6.4, including config file changes and CI updates, are 
maintained on the trunk. This also folds in the merge commit from  #187.

Co-authored-by: Dev Ojha <ValarDragon@users.noreply.github.com>