internal/libs/autofile: continuously .Sync instead of opening and closing every second
This change modifies autofile to no longer close and open the same file
every 1 second but instead to open a single file and then continuously
invoke .Sync, and if for example the file no longer exists, it'll open
again. The premise for this change is that the prior requirements
were built to allow for easy log rotation, but also durability thus
wrote the same file to disk, closed it then reopened it every 1 second,
but unfortunately that consumes a whole lot of file descriptors which
on many systems aren't recycled fast enough.
The prior assumption was also fallacious and unrealistic because most
log rotation happens periodically say 1 hour, or when a file/directory
has reached a threshold of which that would take time and not in 1 second durations.

This conservative approach shows good results with the number of
files opened and closed in my custom instrumented Go build, where I
track and count open vs closed files every 2 seconds.

Once we combine this change with https://github.com/tendermint/tendermint/pull/7327,
it ensures that there is only .Sync on a need to basis and avoids
burning unnecessary resources.

Thanks to Tharsis who paid attention to their customers' reports
and then commissioned me to help resolve these problems.

Fixes #7332