rpc: simplify the JSON-RPC client Caller interface
The old interface returned the result as well as populating it.  Nothing was
using this, so drop the duplicated value from the return signature. Clarify the
documentation on the Caller type.

Update usage of the Caller interface. This has some pass-along effects on some
of the other unexported decoder plumbing. None of the callers was using the
result return value, so this was otherwise mostly cosmetic.

Rework the MapToRequest helper to take an arbitrary value instead of only a
map. This is groundwork for getting rid of the custom marshaling code. For now,
however, the implementation preserves the existing behaviour for the map, until
we can replace those.