Wait for the autofile group to finish on WAL shutdown.
The base WAL wrapper supposedly waits for the autofile group, but only if it
does not see the group already stopped. If the OnStop hook has already been
hit, this wait will not take place, and OnStop doesn't wait on its own.
I suspect this may lead to WAL cleanup racing with test exit.