protoio: fix incorrect test assertion (#7606)
After writing and then reading a bunch of random messages, the test was
checking that it did not read the same number of messages that it wrote.
The sense of this check was inverted; they should match.

Introduced by accident in #7522. I'm not sure why this did not show up in CI.

Edit: I now know why it didn't show up in ci: #7608.
(cherry picked from commit c24f003b556bb5195540337ee9ceb8fef5a0f776)

# Conflicts:
#	internal/libs/protoio/io_test.go