types: rename and extend the EventData interface
This is the interface shared by types that can be used as event data in, for
example, subscriptions via the RPC.

To be compatible with the RPC service, data need to support JSON encoding.
Require this as part of the interface.

Fully propagating this change will require updating the method signatures for
the pubsub package too. For now, to keep this change more self-contained, I
have inserted type assertions with error checks in the RPC subscription
handlers.